requirements.txt
asgiref==3.6.0
Django==4.2
django-environ==0.10.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3


mysqlclient
# PyJWT==2.6.0